requirements.txt
django==2.2
mysqlclient
pillow
django-pure-pagination
django-import-export
redis